Avoid polluting the Task interface
