+++ indexing-service/src/main/java/org/apache/druid/indexing/common/task/IndexTaskUtils.java
@@ -112,7 +112,14 @@ public class IndexTaskUtils
     metricBuilder.setDimension(DruidMetrics.TASK_ID, task.getId());
     metricBuilder.setDimension(DruidMetrics.TASK_TYPE, task.getType());
     metricBuilder.setDimension(DruidMetrics.DATASOURCE, task.getDataSource());
-    metricBuilder.setDimension(DruidMetrics.TASK_INGESTION_MODE, task.getIngestionMode());
+  }
+
+  public static void setTaskDimensions(final ServiceMetricEvent.Builder metricBuilder, final AbstractTask task)
+  {
+    metricBuilder.setDimension(DruidMetrics.TASK_ID, task.getId());
+    metricBuilder.setDimension(DruidMetrics.TASK_TYPE, task.getType());
+    metricBuilder.setDimension(DruidMetrics.DATASOURCE, task.getDataSource());
+    metricBuilder.setDimension(DruidMetrics.TASK_INGESTION_MODE, ((AbstractTask) task).getIngestionMode());
   }
 
   public static void setTaskStatusDimensions(
